Organize examples shared between unit and e2e tests
