--- test/e2e/_examples.cjs
@@ -1,13 +1,0 @@
-/**
- * @overview Provides test examples for end-to-end testing Shescape.
- * @license Unlicense
- * @author Eric Cornelissen <ericornelissen@gmail.com>
- */
-
-const unixExamples = require("../unit/unix/_examples.cjs");
-const winExamples = require("../unit/win/_examples.cjs");
-
-module.exports.unixEscape = unixExamples.escape;
-module.exports.unixQuote = unixExamples.quote;
-module.exports.winEscape = winExamples.escape;
-module.exports.winQuote = winExamples.quote;
