+++ test/e2e/cjs/quote.test.cjs
@@ -6,7 +6,7 @@
 
 const test = require("ava");
 
-const macros = require("../_macros.cjs");
+const macros = require("../macros.cjs");
 
 const { quote } = require("../../../index.cjs");
