+++ test/e2e/esm/escape.test.js
@@ -6,7 +6,7 @@
 
 import test from "ava";
 
-import * as macros from "../_macros.cjs";
+import * as macros from "../macros.cjs";
 
 import { escape } from "../../../index.js";
