+++ test/unit/unix/_fixtures.js
@@ -1,0 +1,7 @@
+/**
+ * @overview Provides fixtures for unit testing Unix functionality
+ * @license Unlicense
+ * @author Eric Cornelissen <ericornelissen@gmail.com>
+ */
+
+export * from "../../fixtures/unix.cjs";
