+++ test/unit/win/_fixtures.js
@@ -1,0 +1,7 @@
+/**
+ * @overview Provides fixtures for unit testing Windows functionality
+ * @license Unlicense
+ * @author Eric Cornelissen <ericornelissen@gmail.com>
+ */
+
+export * from "../../fixtures/win.cjs";
