+++ test/unit/win/quote.test.js
@@ -7,12 +7,12 @@
 
 import test from "ava";
 
-import * as examples from "./_examples.cjs";
+import * as fixtures from "./_fixtures.js";
 import * as macros from "./_macros.js";
 
 import * as win from "../../../src/win.js";
 
-Object.entries(examples.quote).forEach(([shellName, scenarios]) => {
+Object.entries(fixtures.quote).forEach(([shellName, scenarios]) => {
   const cases = Object.values(scenarios).flat();
 
   cases.forEach(({ input, expected }) => {
